+++ server/prisma/migrations/20260808191702_fv_diagnostico_sin_exportacion/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "FvDiagnostico" ADD VALUE 'SIN_EXPORTACION';